[refactor] [history] 히스토리 모드 리팩토링 REF : #18
